彻底使用主线程调用renderer，修复sharedPtr自动销毁资产导致堆错误的问题？

+++ Source/Engine/Structure/SystemConfig.hpp
@@ -5,5 +5,5 @@
 struct SystemConfig
 {
     int TargetFps;
-    Vec2<double> ViewportSize;
+    Vec2<int> ViewportSize;
 };
